LPD-58182 Inline StringQueryTranslator/StringQueryTranslatorImpl
