--- modules/apps/portal-search-elasticsearch7/portal-search-elasticsearch7-impl/src/main/java/com/liferay/portal/search/elasticsearch7/internal/legacy/query/StringQueryTranslatorImpl.java
@@ -1,34 +1,0 @@
-/**
- * SPDX-FileCopyrightText: (c) 2000 Liferay, Inc. https://liferay.com
- * SPDX-License-Identifier: LGPL-2.1-or-later OR LicenseRef-Liferay-DXP-EULA-2.0.0-2023-06
- */
-
-package com.liferay.portal.search.elasticsearch7.internal.legacy.query;
-
-import com.liferay.portal.kernel.search.generic.StringQuery;
-
-import org.elasticsearch.index.query.QueryBuilder;
-import org.elasticsearch.index.query.QueryBuilders;
-import org.elasticsearch.index.query.QueryStringQueryBuilder;
-
-import org.osgi.service.component.annotations.Component;
-
-/**
- * @author Michael C. Han
- */
-@Component(service = StringQueryTranslator.class)
-public class StringQueryTranslatorImpl implements StringQueryTranslator {
-
-	@Override
-	public QueryBuilder translate(StringQuery stringQuery) {
-		QueryStringQueryBuilder queryStringQueryBuilder =
-			QueryBuilders.queryStringQuery(stringQuery.getQuery());
-
-		if (!stringQuery.isDefaultBoost()) {
-			queryStringQueryBuilder.boost(stringQuery.getBoost());
-		}
-
-		return queryStringQueryBuilder;
-	}
-
-}
